refactor: move db connection logic into global space 2022-08-01 (22:42:%)
