+++ src/users/users.module.ts
@@ -3,7 +3,7 @@ import { InMemoryUsersRepository } from 'src/data-access/in-memory-users.reposit
 import { Module } from '@nestjs/common';
 import { UsersController } from './controllers/users.controller';
 import { UsersService } from './services/users.service';
-import { PostgressService } from './services/pg.sequelize.service';
+import { PostgressService } from '../services/pg.sequelize.service';
 
 @Module({
   controllers: [UsersController],
